+++ docs/change_log.rst
@@ -133,3 +133,7 @@ Series
 - Added columns=None default to get_column() on DataFrame
 - Fix bug in get_slice for empty DataFrames
 - Fix bug in DataFrame append for empty DataFrames
+
+2.1.2 (5/20/17)
+~~~~~~~~~~~~~~~
+- Added delete_all_rows() to DataFrame